+++ components/org.wso2.identity.webhook.common.event.handler/src/test/java/org/wso2/identity/webhook/common/event/handler/internal/handler/LoginEventHookHandlerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+ * Copyright (c) 2024-2025, WSO2 LLC. (http://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.common.event.handler;
+package org.wso2.identity.webhook.common.event.handler.internal.handler;
 
 import org.json.simple.JSONObject;
 import org.mockito.ArgumentCaptor;
@@ -25,10 +25,14 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
+import org.wso2.carbon.identity.application.authentication.framework.AuthenticatorStatus;
+import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
+import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 import org.wso2.carbon.identity.configuration.mgt.core.ConfigurationManager;
 import org.wso2.carbon.identity.configuration.mgt.core.exception.ConfigurationManagementException;
 import org.wso2.carbon.identity.configuration.mgt.core.model.Attribute;
@@ -42,44 +46,52 @@ import org.wso2.identity.event.common.publisher.EventPublisherService;
 import org.wso2.identity.event.common.publisher.model.EventContext;
 import org.wso2.identity.event.common.publisher.model.EventPayload;
 import org.wso2.identity.event.common.publisher.model.SecurityEventTokenPayload;
-import org.wso2.identity.webhook.common.event.handler.api.builder.UserOperationEventPayloadBuilder;
+import org.wso2.identity.webhook.common.event.handler.api.builder.LoginEventPayloadBuilder;
+import org.wso2.identity.webhook.common.event.handler.api.constants.EventSchema;
 import org.wso2.identity.webhook.common.event.handler.api.model.EventData;
 import org.wso2.identity.webhook.common.event.handler.internal.component.EventHookHandlerDataHolder;
 import org.wso2.identity.webhook.common.event.handler.internal.config.EventPublisherConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.config.ResourceConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.constant.Constants;
-import org.wso2.identity.webhook.common.event.handler.internal.handler.UserOperationEventHookHandler;
 import org.wso2.identity.webhook.common.event.handler.internal.util.EventConfigManager;
+import org.wso2.identity.webhook.common.event.handler.internal.util.EventHookHandlerUtils;
 import org.wso2.identity.webhook.common.event.handler.internal.util.PayloadBuilderFactory;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import javax.servlet.http.HttpServletRequest;
+
+import static org.mockito.Answers.CALLS_REAL_METHODS;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.withSettings;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedIdentityTenantUtil;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedServiceURLBuilder;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockIdentityTenantUtil;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockServiceURLBuilder;
 
 /**
- * Unit tests for the UserOperationEventHookHandler class and related classes.
+ * Unit tests for the LoginEventHookHandler class and related classes.
  */
-public class UserOperationEventHookHandlerTest {
+public class LoginEventHookHandlerTest {
 
-    private static final String SAMPLE_EVENT_KEY =
-            "schemas.identity.wso2.org/events/user-operations/event-type/updateUserGroup";
+    private static final String SAMPLE_TENANT_DOMAIN = "myorg";
+    private static final String SAMPLE_USER = "user";
+    private static final String SAMPLE_EVENT_KEY_LOGIN_SUCCESS =
+            "schemas.identity.wso2.org/events/logins/event-type/loginSuccess";
+    private static final String SAMPLE_EVENT_KEY_LOGIN_FAILED =
+            "schemas.identity.wso2.org/events/logins/event-type/loginFailed";
     private static final String SAMPLE_ATTRIBUTE_JSON = "{\"sendCredentials\":false,\"publishEnabled\":true}";
-    private static final String DOMAIN_QUALIFIED_ADDED_USER_NAME = "PRIMARY/john";
-    private static final String CARBON_SUPER = "carbon.super";
-    private static final String ADMIN = "ADMIN";
-
     @Mock
     private ConfigurationManager mockedConfigurationManager;
     @Mock
@@ -87,9 +99,11 @@ public class UserOperationEventHookHandlerTest {
     @Mock
     private EventPayload mockedEventPayload;
     @InjectMocks
-    private UserOperationEventHookHandler userOperationEventHookHandler;
+    private LoginEventHookHandler loginEventHookHandler;
     @Mock
-    private UserOperationEventPayloadBuilder mockedUserOperationEventPayloadBuilder;
+    private LoginEventPayloadBuilder mockedLoginEventPayloadBuilder;
+    @Mock
+    private EventHookHandlerUtils mockedEventHookHandlerUtils;
     @Mock
     private EventConfigManager mockedEventConfigManager;
 
@@ -105,16 +119,22 @@ public class UserOperationEventHookHandlerTest {
     @AfterMethod
     public void tearDownMethod() {
 
+        Mockito.reset(mockedEventHookHandlerUtils);
         Mockito.reset(mockedEventPublisherService);
     }
 
     @Test
     public void testCanHandle() {
 
-        Event event = createEvent(IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE);
+        Event event = createEvent(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
         IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
-        boolean canHandle = userOperationEventHookHandler.canHandle(messageContext);
-        assertTrue(canHandle, "The event handler should be able to handle the event POST_UPDATE_USER_LIST_OF_ROLE.");
+        boolean canHandle = loginEventHookHandler.canHandle(messageContext);
+        assertTrue(canHandle, "The event handler should be able to handle the event.");
+
+        event = createEvent(IdentityEventConstants.EventName.AUTHENTICATION_FAILURE.name());
+        messageContext = new IdentityEventMessageContext(event);
+        canHandle = loginEventHookHandler.canHandle(messageContext);
+        assertTrue(canHandle, "Handler should handle AUTHENTICATION_FAILURE event.");
     }
 
     @Test
@@ -122,7 +142,7 @@ public class UserOperationEventHookHandlerTest {
 
         Event event = createEvent(IdentityEventConstants.EventName.SESSION_TERMINATE.name());
         IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
-        boolean canHandle = userOperationEventHookHandler.canHandle(messageContext);
+        boolean canHandle = loginEventHookHandler.canHandle(messageContext);
         assertFalse(canHandle, "The event handler should not be able to handle the SESSION_TERMINATE event.");
     }
 
@@ -130,8 +150,10 @@ public class UserOperationEventHookHandlerTest {
     public Object[][] eventDataProvider() {
 
         return new Object[][]{
-                {IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE,
-                        Constants.EventHandlerKey.WSO2.POST_UPDATE_USER_LIST_OF_ROLE_EVENT, SAMPLE_EVENT_KEY}
+                {IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name(),
+                        Constants.EventHandlerKey.WSO2.LOGIN_SUCCESS_EVENT, SAMPLE_EVENT_KEY_LOGIN_SUCCESS},
+                {IdentityEventConstants.EventName.AUTHENTICATION_STEP_FAILURE.name(),
+                        Constants.EventHandlerKey.WSO2.LOGIN_FAILED_EVENT, SAMPLE_EVENT_KEY_LOGIN_FAILED}
         };
     }
 
@@ -145,16 +167,14 @@ public class UserOperationEventHookHandlerTest {
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getUserOperationEventPayloadBuilder(anyString()))
-                    .thenReturn(mockedUserOperationEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getLoginEventPayloadBuilder(EventSchema.WSO2))
+                    .thenReturn(mockedLoginEventPayloadBuilder);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
             when(mockedEventConfigManager.getEventUri(anyString())).thenReturn(expectedEventKey);
-            when(mockedEventConfigManager.getEventPublisherConfigForTenant(anyString(), anyString())).thenReturn(
-                    eventPublisherConfig);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            userOperationEventHookHandler.handleEvent(event);
+            loginEventHookHandler.handleEvent(event);
 
             verifyEventPublishedWithExpectedKey(expectedEventKey);
         }
@@ -164,26 +184,36 @@ public class UserOperationEventHookHandlerTest {
     public void testHandleEventWithPublishingDisabled() throws
             ConfigurationManagementException, IdentityEventException {
 
-        Event event = createEventWithProperties(IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE);
-        Resources resources =
-                createResourcesWithAttributes(Constants.EventHandlerKey.WSO2.POST_UPDATE_USER_LIST_OF_ROLE_EVENT);
+        Event event = createEventWithProperties(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
+        Resources resources = createResourcesWithAttributes(Constants.EventHandlerKey.WSO2.LOGIN_SUCCESS_EVENT);
         EventPublisherConfig eventPublisherConfig = new EventPublisherConfig(false,
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getUserOperationEventPayloadBuilder(anyString()))
-                    .thenReturn(mockedUserOperationEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getLoginEventPayloadBuilder(EventSchema.WSO2))
+                    .thenReturn(mockedLoginEventPayloadBuilder);
             reset(mockedConfigurationManager);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
-            when(mockedEventConfigManager.getEventPublisherConfigForTenant(anyString(), anyString())).thenReturn(
-                    eventPublisherConfig);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            userOperationEventHookHandler.handleEvent(event);
+            loginEventHookHandler.handleEvent(event);
 
             verify(mockedEventPublisherService, times(0)).publish(any(), any());
         }
+    }
+
+    @Test
+    public void testPassiveAuthenticate() throws IdentityEventException {
+
+        Event event = createEventWithProperties(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
+        AuthenticationContext context = createAuthenticationContext();
+        context.setPassiveAuthenticate(true);
+        event.getEventProperties().put("context", context);
+
+        loginEventHookHandler.handleEvent(event);
+
+        verify(mockedEventPublisherService, times(0)).publish(any(), any());
     }
 
     private void setupDataHolderMocks() {
@@ -194,8 +224,10 @@ public class UserOperationEventHookHandlerTest {
 
     private void setupPayloadBuilderMocks() throws IdentityEventException {
 
-        when(mockedUserOperationEventPayloadBuilder.getEventSchemaType()).thenReturn("WSO2");
-        when(mockedUserOperationEventPayloadBuilder.buildUserGroupUpdateEvent(any(EventData.class)))
+        when(mockedLoginEventPayloadBuilder.getEventSchemaType()).thenReturn(EventSchema.WSO2);
+        when(mockedLoginEventPayloadBuilder.buildAuthenticationSuccessEvent(any(EventData.class)))
+                .thenReturn(mockedEventPayload);
+        when(mockedLoginEventPayloadBuilder.buildAuthenticationFailedEvent(any(EventData.class)))
                 .thenReturn(mockedEventPayload);
     }
 
@@ -203,7 +235,16 @@ public class UserOperationEventHookHandlerTest {
 
         mockServiceURLBuilder();
         mockIdentityTenantUtil();
-        userOperationEventHookHandler = new UserOperationEventHookHandler(mockedEventConfigManager);
+        mockedEventHookHandlerUtils = mock(EventHookHandlerUtils.class, withSettings()
+                .defaultAnswer(CALLS_REAL_METHODS));
+        loginEventHookHandler = new LoginEventHookHandler(mockedEventConfigManager);
+    }
+
+    @AfterClass
+    public void tearDown() {
+
+        closeMockedServiceURLBuilder();
+        closeMockedIdentityTenantUtil();
     }
 
     private Event createEvent(String eventName) {
@@ -214,11 +255,13 @@ public class UserOperationEventHookHandlerTest {
     private Event createEventWithProperties(String eventName) {
 
         HashMap<String, Object> properties = new HashMap<>();
-
-        String[] addedUsers = new String[]{DOMAIN_QUALIFIED_ADDED_USER_NAME};
-        properties.put(IdentityEventConstants.EventProperty.NEW_USERS, addedUsers);
-        properties.put(IdentityEventConstants.EventProperty.TENANT_DOMAIN, CARBON_SUPER);
-        properties.put(IdentityEventConstants.EventProperty.INITIATOR_TYPE, ADMIN);
+        HashMap<String, Object> params = new HashMap<>();
+        AuthenticationContext context = createAuthenticationContext();
+        params.put("request", mock(HttpServletRequest.class));
+        params.put("user", mockAuthenticatedUser());
+        properties.put("context", context);
+        properties.put("authenticationStatus", AuthenticatorStatus.PASS);
+        properties.put("params", params);
 
         return new Event(eventName, properties);
     }
@@ -236,6 +279,20 @@ public class UserOperationEventHookHandlerTest {
         resourceList.add(resource);
         resources.setResources(resourceList);
         return resources;
+    }
+
+    private AuthenticationContext createAuthenticationContext() {
+
+        AuthenticationContext context = new AuthenticationContext();
+        context.setTenantDomain(SAMPLE_TENANT_DOMAIN);
+        return context;
+    }
+
+    private AuthenticatedUser mockAuthenticatedUser() {
+
+        AuthenticatedUser user = new AuthenticatedUser();
+        user.setUserName(SAMPLE_USER);
+        return user;
     }
 
     private void verifyEventPublishedWithExpectedKey(String expectedEventKey) {
