+++ components/org.wso2.identity.webhook.common.event.handler/src/test/java/org/wso2/identity/webhook/common/event/handler/internal/handler/SessionEventHookHandlerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025, WSO2 LLC. (http://www.wso2.com).
+ * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.common.event.handler;
+package org.wso2.identity.webhook.common.event.handler.internal.handler;
 
 import org.json.simple.JSONObject;
 import org.mockito.ArgumentCaptor;
@@ -25,12 +25,14 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import org.wso2.carbon.identity.application.authentication.framework.AuthenticatorStatus;
 import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
+import org.wso2.carbon.identity.application.authentication.framework.context.SessionContext;
 import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 import org.wso2.carbon.identity.configuration.mgt.core.ConfigurationManager;
 import org.wso2.carbon.identity.configuration.mgt.core.exception.ConfigurationManagementException;
@@ -39,26 +41,25 @@ import org.wso2.carbon.identity.configuration.mgt.core.model.Resource;
 import org.wso2.carbon.identity.configuration.mgt.core.model.Resources;
 import org.wso2.carbon.identity.event.IdentityEventConstants;
 import org.wso2.carbon.identity.event.IdentityEventException;
-import org.wso2.carbon.identity.event.bean.IdentityEventMessageContext;
 import org.wso2.carbon.identity.event.event.Event;
 import org.wso2.identity.event.common.publisher.EventPublisherService;
 import org.wso2.identity.event.common.publisher.model.EventContext;
 import org.wso2.identity.event.common.publisher.model.EventPayload;
 import org.wso2.identity.event.common.publisher.model.SecurityEventTokenPayload;
-import org.wso2.identity.webhook.common.event.handler.api.builder.LoginEventPayloadBuilder;
+import org.wso2.identity.webhook.common.event.handler.api.builder.SessionEventPayloadBuilder;
 import org.wso2.identity.webhook.common.event.handler.api.constants.EventSchema;
 import org.wso2.identity.webhook.common.event.handler.api.model.EventData;
 import org.wso2.identity.webhook.common.event.handler.internal.component.EventHookHandlerDataHolder;
 import org.wso2.identity.webhook.common.event.handler.internal.config.EventPublisherConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.config.ResourceConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.constant.Constants;
-import org.wso2.identity.webhook.common.event.handler.internal.handler.LoginEventHookHandler;
 import org.wso2.identity.webhook.common.event.handler.internal.util.EventConfigManager;
 import org.wso2.identity.webhook.common.event.handler.internal.util.EventHookHandlerUtils;
 import org.wso2.identity.webhook.common.event.handler.internal.util.PayloadBuilderFactory;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -73,25 +74,27 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.Mockito.withSettings;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
+import static org.wso2.identity.webhook.common.event.handler.internal.constant.Constants.SP_TO_CARBON_CLAIM_MAPPING;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedIdentityTenantUtil;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedServiceURLBuilder;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockIdentityTenantUtil;
 import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockServiceURLBuilder;
 
 /**
- * Unit tests for the LoginEventHookHandler class and related classes.
+ * Unit tests for the SessionEventHookHandlerTest class and related classes.
  */
-public class LoginEventHookHandlerTest {
+public class SessionEventHookHandlerTest {
 
     private static final String SAMPLE_TENANT_DOMAIN = "myorg";
     private static final String SAMPLE_USER = "user";
-    private static final String SAMPLE_EVENT_KEY_LOGIN_SUCCESS =
-            "schemas.identity.wso2.org/events/logins/event-type/loginSuccess";
-    private static final String SAMPLE_EVENT_KEY_LOGIN_FAILED =
-            "schemas.identity.wso2.org/events/logins/event-type/loginFailed";
+    private static final String SAMPLE_EVENT_KEY_SESSION_REVOKED =
+            "https://schemas.openid.net/secevent/caep/event-type/session-revoked";
+    private static final String SAMPLE_EVENT_KEY_SESSION_ESTABLISHED =
+            "https://schemas.openid.net/secevent/caep/event-type/session-established";
+    private static final String SAMPLE_EVENT_KEY_SESSION_PRESENTED =
+            "https://schemas.openid.net/secevent/caep/event-type/session-presented";
     private static final String SAMPLE_ATTRIBUTE_JSON = "{\"sendCredentials\":false,\"publishEnabled\":true}";
+
     @Mock
     private ConfigurationManager mockedConfigurationManager;
     @Mock
@@ -99,9 +102,9 @@ public class LoginEventHookHandlerTest {
     @Mock
     private EventPayload mockedEventPayload;
     @InjectMocks
-    private LoginEventHookHandler loginEventHookHandler;
+    private SessionEventHookHandler sessionEventHookHandler;
     @Mock
-    private LoginEventPayloadBuilder mockedLoginEventPayloadBuilder;
+    private SessionEventPayloadBuilder mockedSessionEventPayloadBuilder;
     @Mock
     private EventHookHandlerUtils mockedEventHookHandlerUtils;
     @Mock
@@ -116,6 +119,13 @@ public class LoginEventHookHandlerTest {
         setupUtilities();
     }
 
+    @AfterClass
+    public void tearDown() {
+
+        closeMockedServiceURLBuilder();
+        closeMockedIdentityTenantUtil();
+    }
+
     @AfterMethod
     public void tearDownMethod() {
 
@@ -123,38 +133,39 @@ public class LoginEventHookHandlerTest {
         Mockito.reset(mockedEventPublisherService);
     }
 
-    @Test
-    public void testCanHandle() {
-
-        Event event = createEvent(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
-        IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
-        boolean canHandle = loginEventHookHandler.canHandle(messageContext);
-        assertTrue(canHandle, "The event handler should be able to handle the event.");
-
-        event = createEvent(IdentityEventConstants.EventName.AUTHENTICATION_FAILURE.name());
-        messageContext = new IdentityEventMessageContext(event);
-        canHandle = loginEventHookHandler.canHandle(messageContext);
-        assertTrue(canHandle, "Handler should handle AUTHENTICATION_FAILURE event.");
-    }
-
-    @Test
-    public void testCannotHandle() {
-
-        Event event = createEvent(IdentityEventConstants.EventName.SESSION_TERMINATE.name());
-        IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
-        boolean canHandle = loginEventHookHandler.canHandle(messageContext);
-        assertFalse(canHandle, "The event handler should not be able to handle the SESSION_TERMINATE event.");
-    }
-
     @DataProvider(name = "eventDataProvider")
     public Object[][] eventDataProvider() {
 
         return new Object[][]{
-                {IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name(),
-                        Constants.EventHandlerKey.WSO2.LOGIN_SUCCESS_EVENT, SAMPLE_EVENT_KEY_LOGIN_SUCCESS},
-                {IdentityEventConstants.EventName.AUTHENTICATION_STEP_FAILURE.name(),
-                        Constants.EventHandlerKey.WSO2.LOGIN_FAILED_EVENT, SAMPLE_EVENT_KEY_LOGIN_FAILED}
+                {IdentityEventConstants.EventName.SESSION_TERMINATE.name(),
+                        Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT,
+                        SAMPLE_EVENT_KEY_SESSION_REVOKED
+                },
+                {IdentityEventConstants.EventName.SESSION_CREATE.name(),
+                        Constants.EventHandlerKey.CAEP.SESSION_ESTABLISHED_EVENT,
+                        SAMPLE_EVENT_KEY_SESSION_ESTABLISHED
+                },
+                {IdentityEventConstants.EventName.SESSION_UPDATE.name(),
+                        Constants.EventHandlerKey.CAEP.SESSION_PRESENTED_EVENT,
+                        SAMPLE_EVENT_KEY_SESSION_PRESENTED
+                },
+                {IdentityEventConstants.EventName.SESSION_EXPIRE.name(),
+                        Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT,
+                        SAMPLE_EVENT_KEY_SESSION_REVOKED
+                },
+                {IdentityEventConstants.EventName.SESSION_EXTEND.name(),
+                        Constants.EventHandlerKey.CAEP.SESSION_PRESENTED_EVENT,
+                        SAMPLE_EVENT_KEY_SESSION_PRESENTED
+                }
         };
+    }
+
+    @Test
+    public void testGetName() {
+
+        String expectedName = Constants.SESSION_EVENT_HOOK_NAME;
+        String actualName = sessionEventHookHandler.getName();
+        assertEquals(actualName, expectedName, "The name of the event hook handler should match.");
     }
 
     @Test(dataProvider = "eventDataProvider")
@@ -167,14 +178,14 @@ public class LoginEventHookHandlerTest {
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getLoginEventPayloadBuilder(EventSchema.WSO2))
-                    .thenReturn(mockedLoginEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getSessionEventPayloadBuilder(EventSchema.CAEP))
+                    .thenReturn(mockedSessionEventPayloadBuilder);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
             when(mockedEventConfigManager.getEventUri(anyString())).thenReturn(expectedEventKey);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            loginEventHookHandler.handleEvent(event);
+            sessionEventHookHandler.handleEvent(event);
 
             verifyEventPublishedWithExpectedKey(expectedEventKey);
         }
@@ -184,36 +195,23 @@ public class LoginEventHookHandlerTest {
     public void testHandleEventWithPublishingDisabled() throws
             ConfigurationManagementException, IdentityEventException {
 
-        Event event = createEventWithProperties(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
-        Resources resources = createResourcesWithAttributes(Constants.EventHandlerKey.WSO2.LOGIN_SUCCESS_EVENT);
+        Event event = createEventWithProperties(IdentityEventConstants.EventName.SESSION_TERMINATE.name());
+        Resources resources = createResourcesWithAttributes(Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT);
         EventPublisherConfig eventPublisherConfig = new EventPublisherConfig(false,
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getLoginEventPayloadBuilder(EventSchema.WSO2))
-                    .thenReturn(mockedLoginEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getSessionEventPayloadBuilder(EventSchema.CAEP))
+                    .thenReturn(mockedSessionEventPayloadBuilder);
             reset(mockedConfigurationManager);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            loginEventHookHandler.handleEvent(event);
+            sessionEventHookHandler.handleEvent(event);
 
             verify(mockedEventPublisherService, times(0)).publish(any(), any());
         }
-    }
-
-    @Test
-    public void testPassiveAuthenticate() throws IdentityEventException {
-
-        Event event = createEventWithProperties(IdentityEventConstants.EventName.AUTHENTICATION_SUCCESS.name());
-        AuthenticationContext context = createAuthenticationContext();
-        context.setPassiveAuthenticate(true);
-        event.getEventProperties().put("context", context);
-
-        loginEventHookHandler.handleEvent(event);
-
-        verify(mockedEventPublisherService, times(0)).publish(any(), any());
     }
 
     private void setupDataHolderMocks() {
@@ -224,10 +222,16 @@ public class LoginEventHookHandlerTest {
 
     private void setupPayloadBuilderMocks() throws IdentityEventException {
 
-        when(mockedLoginEventPayloadBuilder.getEventSchemaType()).thenReturn(EventSchema.WSO2);
-        when(mockedLoginEventPayloadBuilder.buildAuthenticationSuccessEvent(any(EventData.class)))
+        when(mockedSessionEventPayloadBuilder.getEventSchemaType()).thenReturn(EventSchema.CAEP);
+        when(mockedSessionEventPayloadBuilder.buildSessionUpdateEvent(any(EventData.class)))
                 .thenReturn(mockedEventPayload);
-        when(mockedLoginEventPayloadBuilder.buildAuthenticationFailedEvent(any(EventData.class)))
+        when(mockedSessionEventPayloadBuilder.buildSessionExtendEvent(any(EventData.class)))
+                .thenReturn(mockedEventPayload);
+        when(mockedSessionEventPayloadBuilder.buildSessionTerminateEvent(any(EventData.class)))
+                .thenReturn(mockedEventPayload);
+        when(mockedSessionEventPayloadBuilder.buildSessionCreateEvent(any(EventData.class)))
+                .thenReturn(mockedEventPayload);
+        when(mockedSessionEventPayloadBuilder.buildSessionExpireEvent(any(EventData.class)))
                 .thenReturn(mockedEventPayload);
     }
 
@@ -237,18 +241,7 @@ public class LoginEventHookHandlerTest {
         mockIdentityTenantUtil();
         mockedEventHookHandlerUtils = mock(EventHookHandlerUtils.class, withSettings()
                 .defaultAnswer(CALLS_REAL_METHODS));
-        loginEventHookHandler = new LoginEventHookHandler(mockedEventConfigManager);
-    }
-
-    private void tearDownUtilities() {
-
-        closeMockedServiceURLBuilder();
-        closeMockedIdentityTenantUtil();
-    }
-
-    private Event createEvent(String eventName) {
-
-        return new Event(eventName);
+        sessionEventHookHandler = new SessionEventHookHandler(mockedEventConfigManager);
     }
 
     private Event createEventWithProperties(String eventName) {
@@ -256,11 +249,13 @@ public class LoginEventHookHandlerTest {
         HashMap<String, Object> properties = new HashMap<>();
         HashMap<String, Object> params = new HashMap<>();
         AuthenticationContext context = createAuthenticationContext();
+        SessionContext sessionContext = createSessionContext();
         params.put("request", mock(HttpServletRequest.class));
         params.put("user", mockAuthenticatedUser());
         properties.put("context", context);
         properties.put("authenticationStatus", AuthenticatorStatus.PASS);
         properties.put("params", params);
+        properties.put("sessionContext", sessionContext);
 
         return new Event(eventName, properties);
     }
@@ -284,13 +279,33 @@ public class LoginEventHookHandlerTest {
 
         AuthenticationContext context = new AuthenticationContext();
         context.setTenantDomain(SAMPLE_TENANT_DOMAIN);
+        Map<String, String> sampleClaimMapping = new HashMap<>();
+        sampleClaimMapping.put("roles", "http://wso2.org/claims/roles");
+        sampleClaimMapping.put("default_tenant", "http://wso2.org/claims/runtime/default_tenant");
+        sampleClaimMapping.put("active", "http://wso2.org/claims/active");
+        sampleClaimMapping.put("preferred_username", "http://wso2.org/claims/displayName");
+        sampleClaimMapping.put("given_name", "http://wso2.org/claims/givenname");
+        sampleClaimMapping.put("family_name", "http://wso2.org/claims/lastname");
+        sampleClaimMapping.put("email", "http://wso2.org/claims/emailaddress");
+        sampleClaimMapping.put("username", "http://wso2.org/claims/username");
+        sampleClaimMapping.put("associated_tenants", "http://wso2.org/claims/runtime/associated_tenants");
+        context.setProperty(SP_TO_CARBON_CLAIM_MAPPING, sampleClaimMapping);
         return context;
+    }
+
+    private SessionContext createSessionContext() {
+
+        SessionContext sessionContext = new SessionContext();
+        sessionContext.addProperty("tenantDomain", SAMPLE_TENANT_DOMAIN);
+        return sessionContext;
     }
 
     private AuthenticatedUser mockAuthenticatedUser() {
 
         AuthenticatedUser user = new AuthenticatedUser();
         user.setUserName(SAMPLE_USER);
+        user.setTenantDomain(SAMPLE_TENANT_DOMAIN);
+        user.setUserId("123");
         return user;
     }
