+++ components/org.wso2.identity.webhook.common.event.handler/src/test/java/org/wso2/identity/webhook/common/event/handler/internal/handler/UserOperationEventHookHandlerTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.common.event.handler;
+package org.wso2.identity.webhook.common.event.handler.internal.handler;
 
 import org.json.simple.JSONObject;
 import org.mockito.ArgumentCaptor;
@@ -25,14 +25,11 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import org.wso2.carbon.identity.application.authentication.framework.AuthenticatorStatus;
-import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
-import org.wso2.carbon.identity.application.authentication.framework.context.SessionContext;
-import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 import org.wso2.carbon.identity.configuration.mgt.core.ConfigurationManager;
 import org.wso2.carbon.identity.configuration.mgt.core.exception.ConfigurationManagementException;
 import org.wso2.carbon.identity.configuration.mgt.core.model.Attribute;
@@ -40,28 +37,24 @@ import org.wso2.carbon.identity.configuration.mgt.core.model.Resource;
 import org.wso2.carbon.identity.configuration.mgt.core.model.Resources;
 import org.wso2.carbon.identity.event.IdentityEventConstants;
 import org.wso2.carbon.identity.event.IdentityEventException;
+import org.wso2.carbon.identity.event.bean.IdentityEventMessageContext;
 import org.wso2.carbon.identity.event.event.Event;
 import org.wso2.identity.event.common.publisher.EventPublisherService;
 import org.wso2.identity.event.common.publisher.model.EventContext;
 import org.wso2.identity.event.common.publisher.model.EventPayload;
 import org.wso2.identity.event.common.publisher.model.SecurityEventTokenPayload;
-import org.wso2.identity.webhook.common.event.handler.api.builder.SessionEventPayloadBuilder;
-import org.wso2.identity.webhook.common.event.handler.api.constants.EventSchema;
+import org.wso2.identity.webhook.common.event.handler.api.builder.UserOperationEventPayloadBuilder;
 import org.wso2.identity.webhook.common.event.handler.api.model.EventData;
 import org.wso2.identity.webhook.common.event.handler.internal.component.EventHookHandlerDataHolder;
 import org.wso2.identity.webhook.common.event.handler.internal.config.EventPublisherConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.config.ResourceConfig;
 import org.wso2.identity.webhook.common.event.handler.internal.constant.Constants;
-import org.wso2.identity.webhook.common.event.handler.internal.handler.SessionEventHookHandler;
 import org.wso2.identity.webhook.common.event.handler.internal.util.EventConfigManager;
 import org.wso2.identity.webhook.common.event.handler.internal.util.EventHookHandlerUtils;
 import org.wso2.identity.webhook.common.event.handler.internal.util.PayloadBuilderFactory;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
 
 import static org.mockito.Answers.CALLS_REAL_METHODS;
 import static org.mockito.ArgumentMatchers.any;
@@ -74,22 +67,24 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.Mockito.withSettings;
 import static org.testng.Assert.assertEquals;
-import static org.wso2.identity.webhook.common.event.handler.internal.constant.Constants.SP_TO_CARBON_CLAIM_MAPPING;
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertTrue;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedIdentityTenantUtil;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.closeMockedServiceURLBuilder;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockIdentityTenantUtil;
+import static org.wso2.identity.webhook.common.event.handler.util.TestUtils.mockServiceURLBuilder;
 
 /**
- * Unit tests for the SessionEventHookHandlerTest class and related classes.
+ * Unit tests for the UserOperationEventHookHandler class and related classes.
  */
-public class SessionEventHookHandlerTest {
+public class UserOperationEventHookHandlerTest {
 
-    private static final String SAMPLE_TENANT_DOMAIN = "myorg";
-    private static final String SAMPLE_USER = "user";
-    private static final String SAMPLE_EVENT_KEY_SESSION_REVOKED =
-            "https://schemas.openid.net/secevent/caep/event-type/session-revoked";
-    private static final String SAMPLE_EVENT_KEY_SESSION_ESTABLISHED =
-            "https://schemas.openid.net/secevent/caep/event-type/session-established";
-    private static final String SAMPLE_EVENT_KEY_SESSION_PRESENTED =
-            "https://schemas.openid.net/secevent/caep/event-type/session-presented";
+    private static final String SAMPLE_EVENT_KEY =
+            "schemas.identity.wso2.org/events/user-operations/event-type/updateUserGroup";
     private static final String SAMPLE_ATTRIBUTE_JSON = "{\"sendCredentials\":false,\"publishEnabled\":true}";
+    private static final String DOMAIN_QUALIFIED_ADDED_USER_NAME = "PRIMARY/john";
+    private static final String CARBON_SUPER = "carbon.super";
+    private static final String ADMIN = "ADMIN";
 
     @Mock
     private ConfigurationManager mockedConfigurationManager;
@@ -98,11 +93,11 @@ public class SessionEventHookHandlerTest {
     @Mock
     private EventPayload mockedEventPayload;
     @InjectMocks
-    private SessionEventHookHandler sessionEventHookHandler;
-    @Mock
-    private SessionEventPayloadBuilder mockedSessionEventPayloadBuilder;
+    private UserOperationEventHookHandler userOperationEventHookHandler;
     @Mock
     private EventHookHandlerUtils mockedEventHookHandlerUtils;
+    @Mock
+    private UserOperationEventPayloadBuilder mockedUserOperationEventPayloadBuilder;
     @Mock
     private EventConfigManager mockedEventConfigManager;
 
@@ -115,46 +110,44 @@ public class SessionEventHookHandlerTest {
         setupUtilities();
     }
 
+    @AfterClass
+    public void tearDown() {
+
+        closeMockedServiceURLBuilder();
+        closeMockedIdentityTenantUtil();
+    }
+
     @AfterMethod
     public void tearDownMethod() {
-
         Mockito.reset(mockedEventHookHandlerUtils);
         Mockito.reset(mockedEventPublisherService);
+    }
+
+    @Test
+    public void testCanHandle() {
+
+        Event event = createEvent(IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE);
+        IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
+        boolean canHandle = userOperationEventHookHandler.canHandle(messageContext);
+        assertTrue(canHandle, "The event handler should be able to handle the event POST_UPDATE_USER_LIST_OF_ROLE.");
+    }
+
+    @Test
+    public void testCannotHandle() {
+
+        Event event = createEvent(IdentityEventConstants.EventName.SESSION_TERMINATE.name());
+        IdentityEventMessageContext messageContext = new IdentityEventMessageContext(event);
+        boolean canHandle = userOperationEventHookHandler.canHandle(messageContext);
+        assertFalse(canHandle, "The event handler should not be able to handle the SESSION_TERMINATE event.");
     }
 
     @DataProvider(name = "eventDataProvider")
     public Object[][] eventDataProvider() {
 
         return new Object[][]{
-                {IdentityEventConstants.EventName.SESSION_TERMINATE.name(),
-                        Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT,
-                        SAMPLE_EVENT_KEY_SESSION_REVOKED
-                },
-                {IdentityEventConstants.EventName.SESSION_CREATE.name(),
-                        Constants.EventHandlerKey.CAEP.SESSION_ESTABLISHED_EVENT,
-                        SAMPLE_EVENT_KEY_SESSION_ESTABLISHED
-                },
-                {IdentityEventConstants.EventName.SESSION_UPDATE.name(),
-                        Constants.EventHandlerKey.CAEP.SESSION_PRESENTED_EVENT,
-                        SAMPLE_EVENT_KEY_SESSION_PRESENTED
-                },
-                {IdentityEventConstants.EventName.SESSION_EXPIRE.name(),
-                        Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT,
-                        SAMPLE_EVENT_KEY_SESSION_REVOKED
-                },
-                {IdentityEventConstants.EventName.SESSION_EXTEND.name(),
-                        Constants.EventHandlerKey.CAEP.SESSION_PRESENTED_EVENT,
-                        SAMPLE_EVENT_KEY_SESSION_PRESENTED
-                }
+                {IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE,
+                        Constants.EventHandlerKey.WSO2.POST_UPDATE_USER_LIST_OF_ROLE_EVENT, SAMPLE_EVENT_KEY}
         };
-    }
-
-    @Test
-    public void testGetName() {
-
-        String expectedName = Constants.SESSION_EVENT_HOOK_NAME;
-        String actualName = sessionEventHookHandler.getName();
-        assertEquals(actualName, expectedName, "The name of the event hook handler should match.");
     }
 
     @Test(dataProvider = "eventDataProvider")
@@ -167,14 +160,16 @@ public class SessionEventHookHandlerTest {
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getSessionEventPayloadBuilder(EventSchema.CAEP))
-                    .thenReturn(mockedSessionEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getUserOperationEventPayloadBuilder(anyString()))
+                    .thenReturn(mockedUserOperationEventPayloadBuilder);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
             when(mockedEventConfigManager.getEventUri(anyString())).thenReturn(expectedEventKey);
+            when(mockedEventConfigManager.getEventPublisherConfigForTenant(anyString(), anyString())).thenReturn(
+                    eventPublisherConfig);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            sessionEventHookHandler.handleEvent(event);
+            userOperationEventHookHandler.handleEvent(event);
 
             verifyEventPublishedWithExpectedKey(expectedEventKey);
         }
@@ -184,20 +179,23 @@ public class SessionEventHookHandlerTest {
     public void testHandleEventWithPublishingDisabled() throws
             ConfigurationManagementException, IdentityEventException {
 
-        Event event = createEventWithProperties(IdentityEventConstants.EventName.SESSION_TERMINATE.name());
-        Resources resources = createResourcesWithAttributes(Constants.EventHandlerKey.CAEP.SESSION_REVOKED_EVENT);
+        Event event = createEventWithProperties(IdentityEventConstants.Event.POST_UPDATE_USER_LIST_OF_ROLE);
+        Resources resources =
+                createResourcesWithAttributes(Constants.EventHandlerKey.WSO2.POST_UPDATE_USER_LIST_OF_ROLE_EVENT);
         EventPublisherConfig eventPublisherConfig = new EventPublisherConfig(false,
                 new ResourceConfig(new JSONObject()));
 
         try (MockedStatic<PayloadBuilderFactory> mocked = mockStatic(PayloadBuilderFactory.class)) {
-            mocked.when(() -> PayloadBuilderFactory.getSessionEventPayloadBuilder(EventSchema.CAEP))
-                    .thenReturn(mockedSessionEventPayloadBuilder);
+            mocked.when(() -> PayloadBuilderFactory.getUserOperationEventPayloadBuilder(anyString()))
+                    .thenReturn(mockedUserOperationEventPayloadBuilder);
             reset(mockedConfigurationManager);
             when(mockedConfigurationManager.getTenantResources(anyString(), any())).thenReturn(resources);
+            when(mockedEventConfigManager.getEventPublisherConfigForTenant(anyString(), anyString())).thenReturn(
+                    eventPublisherConfig);
             when(mockedEventConfigManager.extractEventPublisherConfig(any(Resources.class), anyString()))
                     .thenReturn(eventPublisherConfig);
 
-            sessionEventHookHandler.handleEvent(event);
+            userOperationEventHookHandler.handleEvent(event);
 
             verify(mockedEventPublisherService, times(0)).publish(any(), any());
         }
@@ -211,38 +209,33 @@ public class SessionEventHookHandlerTest {
 
     private void setupPayloadBuilderMocks() throws IdentityEventException {
 
-        when(mockedSessionEventPayloadBuilder.getEventSchemaType()).thenReturn(EventSchema.CAEP);
-        when(mockedSessionEventPayloadBuilder.buildSessionUpdateEvent(any(EventData.class)))
-                .thenReturn(mockedEventPayload);
-        when(mockedSessionEventPayloadBuilder.buildSessionExtendEvent(any(EventData.class)))
-                .thenReturn(mockedEventPayload);
-        when(mockedSessionEventPayloadBuilder.buildSessionTerminateEvent(any(EventData.class)))
-                .thenReturn(mockedEventPayload);
-        when(mockedSessionEventPayloadBuilder.buildSessionCreateEvent(any(EventData.class)))
-                .thenReturn(mockedEventPayload);
-        when(mockedSessionEventPayloadBuilder.buildSessionExpireEvent(any(EventData.class)))
+        when(mockedUserOperationEventPayloadBuilder.getEventSchemaType()).thenReturn("WSO2");
+        when(mockedUserOperationEventPayloadBuilder.buildUserGroupUpdateEvent(any(EventData.class)))
                 .thenReturn(mockedEventPayload);
     }
 
     private void setupUtilities() {
 
+        mockServiceURLBuilder();
+        mockIdentityTenantUtil();
         mockedEventHookHandlerUtils = mock(EventHookHandlerUtils.class, withSettings()
                 .defaultAnswer(CALLS_REAL_METHODS));
-        sessionEventHookHandler = new SessionEventHookHandler(mockedEventConfigManager);
+        userOperationEventHookHandler = new UserOperationEventHookHandler(mockedEventConfigManager);
+    }
+
+    private Event createEvent(String eventName) {
+
+        return new Event(eventName);
     }
 
     private Event createEventWithProperties(String eventName) {
 
         HashMap<String, Object> properties = new HashMap<>();
-        HashMap<String, Object> params = new HashMap<>();
-        AuthenticationContext context = createAuthenticationContext();
-        SessionContext sessionContext = createSessionContext();
-        params.put("request", mock(HttpServletRequest.class));
-        params.put("user", mockAuthenticatedUser());
-        properties.put("context", context);
-        properties.put("authenticationStatus", AuthenticatorStatus.PASS);
-        properties.put("params", params);
-        properties.put("sessionContext", sessionContext);
+
+        String[] addedUsers = new String[]{DOMAIN_QUALIFIED_ADDED_USER_NAME};
+        properties.put(IdentityEventConstants.EventProperty.NEW_USERS, addedUsers);
+        properties.put(IdentityEventConstants.EventProperty.TENANT_DOMAIN, CARBON_SUPER);
+        properties.put(IdentityEventConstants.EventProperty.INITIATOR_TYPE, ADMIN);
 
         return new Event(eventName, properties);
     }
@@ -260,40 +253,6 @@ public class SessionEventHookHandlerTest {
         resourceList.add(resource);
         resources.setResources(resourceList);
         return resources;
-    }
-
-    private AuthenticationContext createAuthenticationContext() {
-
-        AuthenticationContext context = new AuthenticationContext();
-        context.setTenantDomain(SAMPLE_TENANT_DOMAIN);
-        Map<String, String> sampleClaimMapping = new HashMap<>();
-        sampleClaimMapping.put("roles", "http://wso2.org/claims/roles");
-        sampleClaimMapping.put("default_tenant", "http://wso2.org/claims/runtime/default_tenant");
-        sampleClaimMapping.put("active", "http://wso2.org/claims/active");
-        sampleClaimMapping.put("preferred_username", "http://wso2.org/claims/displayName");
-        sampleClaimMapping.put("given_name", "http://wso2.org/claims/givenname");
-        sampleClaimMapping.put("family_name", "http://wso2.org/claims/lastname");
-        sampleClaimMapping.put("email", "http://wso2.org/claims/emailaddress");
-        sampleClaimMapping.put("username", "http://wso2.org/claims/username");
-        sampleClaimMapping.put("associated_tenants", "http://wso2.org/claims/runtime/associated_tenants");
-        context.setProperty(SP_TO_CARBON_CLAIM_MAPPING, sampleClaimMapping);
-        return context;
-    }
-
-    private SessionContext createSessionContext() {
-
-        SessionContext sessionContext = new SessionContext();
-        sessionContext.addProperty("tenantDomain", SAMPLE_TENANT_DOMAIN);
-        return sessionContext;
-    }
-
-    private AuthenticatedUser mockAuthenticatedUser() {
-
-        AuthenticatedUser user = new AuthenticatedUser();
-        user.setUserName(SAMPLE_USER);
-        user.setTenantDomain(SAMPLE_TENANT_DOMAIN);
-        user.setUserId("123");
-        return user;
     }
 
     private void verifyEventPublishedWithExpectedKey(String expectedEventKey) {
