+++ components/org.wso2.identity.webhook.common.event.handler/src/test/resources/testng.xml
@@ -23,9 +23,10 @@
             <class name="org.wso2.identity.webhook.common.event.handler.util.EventHookHandlerUtilsTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.util.EventPayloadUtilsTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.component.EventHookHandlerServiceComponentTest"/>
-            <class name="org.wso2.identity.webhook.common.event.handler.LoginEventHookHandlerTest"/>
-            <class name="org.wso2.identity.webhook.common.event.handler.SessionEventHookHandlerTest"/>
-            <class name="org.wso2.identity.webhook.common.event.handler.CredentialEventHookHandlerTest"/>
+            <class name="org.wso2.identity.webhook.common.event.handler.internal.handler.LoginEventHookHandlerTest"/>
+            <class name="org.wso2.identity.webhook.common.event.handler.internal.handler.SessionEventHookHandlerTest"/>
+            <class name="org.wso2.identity.webhook.common.event.handler.internal.handler.UserOperationEventHookHandlerTest"/>
+            <class name="org.wso2.identity.webhook.common.event.handler.internal.handler.CredentialEventHookHandlerTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.PayloadBuilderFactoryTest"/>
             <class name="org.wso2.identity.webhook.common.event.handler.component.EventHookHandlerDataHolderTest"/>
         </classes>
